+++ core/src/main/java/org/apache/accumulo/core/spi/scan/DefaultScanInterpreter.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.accumulo.core.util.interpret;
+package org.apache.accumulo.core.spi.scan;
 
 import org.apache.hadoop.io.Text;
