+++ test/src/main/java/org/apache/accumulo/test/ShellIT.java
@@ -24,10 +24,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.file.Files;
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
 import java.util.Arrays;
-import java.util.Date;
 import java.util.List;
 import java.util.TimeZone;
 
@@ -340,27 +337,6 @@ public class ShellIT extends SharedMiniClusterBase {
         expectedFew);
     exec("setauths -c", true);
     exec("deletetable test -f", true, "Table: [test] has been deleted");
-  }
-
-  @Test
-  public void scanDateStringFormatterTest() throws IOException {
-    Shell.log.debug("Starting scan dateStringFormatter test --------------------------");
-    exec("createtable t", true);
-    exec("insert r f q v -ts 0", true);
-    @SuppressWarnings("deprecation")
-    DateFormat dateFormat = new SimpleDateFormat(
-        org.apache.accumulo.core.util.format.DateStringFormatter.DATE_FORMAT);
-    String expected = String.format("r f:q [] %s    v", dateFormat.format(new Date(0)));
-    // historically, showing few did not pertain to ColVis or Timestamp
-    String expectedNoTimestamp = "r f:q []    v";
-    exec("scan -fm org.apache.accumulo.core.util.format.DateStringFormatter -st", true, expected);
-    exec("scan -fm org.apache.accumulo.core.util.format.DateStringFormatter -st -f 1000", true,
-        expected);
-    exec("scan -fm org.apache.accumulo.core.util.format.DateStringFormatter -st -f 5", true,
-        expected);
-    exec("scan -fm org.apache.accumulo.core.util.format.DateStringFormatter", true,
-        expectedNoTimestamp);
-    exec("deletetable t -f", true, "Table: [t] has been deleted");
   }
 
   @Test
